*: Move indexing utilities into graph, keep graph-core db agnostic
Move get_valid_value_type() to graph::data::graphql::validation.
Move attribute_indexing_operations() to graph::data::subgraph::schema.
Build index type and operator in graph-store-postgres to keep graph-core
database agnostic.
