+++ graph/src/data/subgraph/schema.rs
@@ -11,14 +11,17 @@
 //!
 //! See `subgraphs.graphql` in the store for corresponding graphql schema.
 
+use graphql_parser::schema::{Definition, Document, TypeDefinition};
 use hex;
 use rand::rngs::OsRng;
 use rand::Rng;
 use web3::types::*;
+use std::collections::HashMap;
 
 use super::SubgraphDeploymentId;
 use components::ethereum::EthereumBlockPointer;
-use components::store::{EntityFilter, EntityKey, EntityOperation, EntityQuery};
+use components::store::{AttributeIndexOperation, EntityFilter, EntityKey, EntityOperation, EntityQuery};
+use data::graphql::validation::get_valid_value_type;
 use data::store::{Entity, NodeId, SubgraphEntityPair, Value};
 use data::subgraph::{SubgraphManifest, SubgraphName};
 
@@ -620,4 +623,36 @@ pub fn generate_entity_id() -> String {
     // Comparable to uuidv4, but without the hyphens,
     // and without spending bits on a version identifier.
     hex::encode(id_bytes)
+}
+
+pub fn attribute_indexing_operations(
+    subgraph_id: SubgraphDeploymentId,
+    document: Document,
+) -> Vec<AttributeIndexOperation> {
+    let mut indexing_ops = vec![];
+    for (entity_number, schema_type) in document.definitions.clone().into_iter().enumerate() {
+        if let Definition::TypeDefinition(definition) = schema_type {
+            if let TypeDefinition::Object(schema_object) = definition {
+                for (attribute_number, entity_field) in schema_object.fields.into_iter().enumerate()
+                {
+                    indexing_ops.push(AttributeIndexOperation {
+                        subgraph_id: subgraph_id.clone(),
+                        index_name: format!(
+                            "{}_{}_{}_idx",
+                            subgraph_id.clone(),
+                            entity_number,
+                            attribute_number,
+                        ),
+                        field_value_type: match get_valid_value_type(&entity_field.field_type) {
+                            Ok(value_type) => value_type,
+                            Err(_) => continue,
+                        },
+                        attribute_name: entity_field.name,
+                        entity_name: schema_object.name.clone(),
+                    });
+                }
+            }
+        }
+    }
+    indexing_ops
 }
