+++ graphql/src/schema/ast.rs
@@ -161,15 +161,6 @@ pub fn get_field_value_type(field_type: &Type) -> Result<ValueType, Error> {
     }
 }
 
-/// Returns the value type for a GraphQL field type.
-pub fn get_valid_value_type(field_type: &Type) -> Result<ValueType, Error> {
-    match field_type {
-        Type::NamedType(ref name) => ValueType::from_str(&name),
-        Type::NonNullType(inner) => get_valid_value_type(&inner),
-        Type::ListType(inner) => get_valid_value_type(&inner),
-    }
-}
-
 /// Returns the type with the given name.
 pub fn get_named_type<'a>(schema: &'a Document, name: &Name) -> Option<&'a TypeDefinition> {
     schema
