+++ graphql/tests/query.rs
@@ -150,10 +150,7 @@ impl Store for TestStore {
         unimplemented!()
     }
 
-    fn build_entity_attribute_indexes(
-        &self,
-        operations: Vec<AttributeIndexOperation>,
-    ) -> Result<(), Error> {
+    fn build_entity_attribute_indexes(&self, _: Vec<AttributeIndexOperation>) -> Result<(), Error> {
         unimplemented!()
     }
